backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
prisma
pydantic
requests
email-validator
jinja2
httpx
openai